[FIX] account: proper tax repartition for tax lines when using a writeoff reconciliation model with taxes
When creating the results of a writeoff button for the reconciliation widget, debit and credit are actually inverted with regards to what will actually get created. Because of that, the conditions defining whether or not the compute_all should compute refund repartition need to be inverted as well.

Before that, compute_tax_tag_invert computed a different value for the tax_tag_invert field on base and tax line (as the tax line uses the repartition type of tax_repartition_line_id, and the base line depends on debit/credit and the tax type). Hence, the tax_tag_invert values between those two lines where inconsistent, and the tax report was wrong (since we did * -1 on one of them, but not the other).

OPW 2769586

closes odoo/odoo#86091

Signed-off-by: Laurent Smet <las@odoo.com>